src/test/java/com/example/proyectounieventos/servicios/impl/CarritoServiciosImplementTest.java
package com.example.proyectounieventos.servicios.impl;

import static org.junit.jupiter.api.Assertions.*;
import static org.junit.jupiter.api.Assertions.*;
import static org.mockito.ArgumentMatchers.any;

import com.example.proyectounieventos.modelo.documentos.Carrito;
import com.example.proyectounieventos.modelo.documentos.Compra;
import com.example.proyectounieventos.modelo.vo.DetalleCompra;
import com.example.proyectounieventos.repositorios.CarritoRepo;
import org.bson.types.ObjectId;
import org.junit.jupiter.api.Test;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.Mockito;
import org.springframework.boot.test.context.SpringBootTest;

import java.util.ArrayList;
import java.util.List;
import java.util.Optional;

// ... otras importaciones

@SpringBootTest
class CarritoServiciosImplementTest {

    @InjectMocks
    private CarritoServiciosImplement carritoServicios;

    @Mock
    private CarritoRepo carritoRepo;

    // ...

    @Test
    void agregarItem_shouldAddDetalleCompraToCarrito() {
        // ... (ya lo hemos visto)
    }

    @Test
    void actualizarValorTotal_shouldUpdateTotalCorrectly() throws Exception {
        // Configuración
        Carrito carrito = new Carrito();
        // ... agregar items al carrito con diferentes valores
        Mockito.when(carritoRepo.findByIdUsuario(any())).thenReturn(Optional.of(carrito));

        // Ejecución
        carritoServicios.actualizarValorTotal(ObjectId.get());

        // Verificación
        // Asegurarse que el valor total del carrito sea la suma de los valores de los items
    }

    // ...
    @Test
    void eliminarItem_shouldRemoveItemFromCarrito() throws Exception {
        // Configuración

        Carrito carrito = new Carrito(); // Crea una instancia de Carrito
        carrito.setItems(new ArrayList<>()); // Inicializa la lista de items
        DetalleCompra detalleCompra = new DetalleCompra();
         // Asegúrate de que items esté inicializado
        Compra compra = new Compra(List.of(detalleCompra));
        carrito.getItems().add(compra);
        Mockito.when(carritoRepo.findByIdUsuario(any())).thenReturn(Optional.of(carrito));

        // Ejecución
        carritoServicios.eliminarItem(detalleCompra);

        // Verificación
        assertTrue(carrito.getItems().get(0).getDetalles().isEmpty());
    }

    @Test
    void vaciarCarrito_shouldEmptyCarrito() throws Exception {
        // Configuración
        Carrito carrito = new Carrito();
        carrito.getItems().add(new Compra());
        Mockito.when(carritoRepo.findByIdUsuario(any())).thenReturn(Optional.of(carrito));

        // Ejecución
        carritoServicios.vaciarCarrito();

        // Verificación
        assertTrue(carrito.getItems().isEmpty());
    }

    @Test
    void verItems_shouldReturnListOfDetalles() throws Exception {
        // Configuración
        Carrito carrito = new Carrito();
        Compra compra = new Compra();
        compra.setDetalles(List.of(new DetalleCompra(), new DetalleCompra()));
        carrito.getItems().add(compra);
        Mockito.when(carritoRepo.findByIdUsuario(any())).thenReturn(Optional.of(carrito));

        // Ejecución
        List<DetalleCompra> detalles = carritoServicios.verItems(ObjectId.get());

        // Verificación
        int expectedSize = 2; // Ajusta según tus datos de prueba
        assertEquals(expectedSize, detalles.size());
    }

    @Test
    void agregarItem_shouldThrowExceptionWhenDetalleCompraIsNull() {
        // ...
    }

    // ... otras pruebas para diferentes escenarios y excepciones
}